pull data from running task thread

+++ EvolutionaryEngine/src/engine/models/EngineProgressInterface.java
@@ -1,5 +1,5 @@
 package engine.models;
 
 public interface EngineProgressInterface  {
-   void update(int work,int done);
+   void update(double work,double done);
 }
